Fix #2705: Scaffolding failure on large buildings

+++ src/OpenLoco/src/Paint/PaintBuilding.cpp
@@ -73,7 +73,7 @@ namespace OpenLoco::Paint
             }
             ImageId scaffImage = baseScaffImage.withIndexOffset(scaffImages.back);
             auto segmentImageOffset = imageOffset;
-            for (int8_t remainingHeight = totalSectionHeight; remainingHeight > 0; remainingHeight -= segmentHeight, segmentImageOffset.z += segmentHeight)
+            for (int16_t remainingHeight = totalSectionHeight; remainingHeight > 0; remainingHeight -= segmentHeight, segmentImageOffset.z += segmentHeight)
             {
                 session.addToPlotListAsChild(scaffImage, segmentImageOffset, bbOffset, bbSize);
             }
@@ -129,7 +129,7 @@ namespace OpenLoco::Paint
             }
             auto scaffImage = baseScaffImage.withIndexOffset(scaffImages.front);
             auto segmentImageOffset = imageOffset;
-            for (int8_t remainingHeight = totalSectionHeight; remainingHeight > 0; remainingHeight -= segmentHeight, segmentImageOffset.z += segmentHeight)
+            for (int16_t remainingHeight = totalSectionHeight; remainingHeight > 0; remainingHeight -= segmentHeight, segmentImageOffset.z += segmentHeight)
             {
                 session.addToPlotListAsChild(scaffImage, segmentImageOffset, bbOffset, bbSize);
             }
